Rename openshift directory to kubernetes
