+++ testsuite/backend/httpbin.py
@@ -3,10 +3,10 @@
 from functools import cached_property
 
 from testsuite.backend import Backend
-from testsuite.openshift import Selector
-from testsuite.openshift.client import OpenShiftClient
-from testsuite.openshift.deployment import Deployment
-from testsuite.openshift.service import Service, ServicePort
+from testsuite.kubernetes import Selector
+from testsuite.kubernetes.client import OpenShiftClient
+from testsuite.kubernetes.deployment import Deployment
+from testsuite.kubernetes.service import Service, ServicePort
 
 
 class Httpbin(Backend):
