+++ testsuite/gateway/gateway_api/route.py
@@ -6,8 +6,8 @@ from httpx import Client
 
 from testsuite.httpx import KuadrantClient
 from testsuite.gateway import Gateway, GatewayRoute, PathMatch, MatchType, RouteMatch
-from testsuite.openshift.client import OpenShiftClient
-from testsuite.openshift import OpenShiftObject, modify
+from testsuite.kubernetes.client import OpenShiftClient
+from testsuite.kubernetes import OpenShiftObject, modify
 from testsuite.policy import Policy
 from testsuite.utils import asdict, check_condition
