+++ testsuite/kubernetes/limitador.py
@@ -2,8 +2,8 @@
 
 from openshift_client import selector
 
-from testsuite.openshift import CustomResource
-from testsuite.openshift.deployment import Deployment
+from testsuite.kubernetes import CustomResource
+from testsuite.kubernetes.deployment import Deployment
 
 
 class LimitadorCR(CustomResource):
