+++ testsuite/tests/kuadrant/authorino/operator/raw_http/conftest.py
@@ -6,7 +6,7 @@ from testsuite.gateway.exposers import LoadBalancerServiceExposer
 from testsuite.policy.authorization import Value, JsonResponse
 from testsuite.httpx import KuadrantClient
 from testsuite.policy.authorization.auth_config import AuthConfig
-from testsuite.openshift.route import OpenshiftRoute
+from testsuite.kubernetes.route import OpenshiftRoute
 
 
 @pytest.fixture(scope="module")
