+++ testsuite/tests/kuadrant/conftest.py
@@ -10,7 +10,7 @@ from testsuite.gateway.envoy import Envoy
 from testsuite.gateway.envoy.route import EnvoyVirtualRoute
 from testsuite.gateway.gateway_api.gateway import KuadrantGateway
 from testsuite.gateway.gateway_api.route import HTTPRoute
-from testsuite.openshift.kuadrant import KuadrantCR
+from testsuite.kubernetes.kuadrant import KuadrantCR
 from testsuite.policy.authorization.auth_policy import AuthPolicy
 from testsuite.policy.rate_limit_policy import RateLimitPolicy
